moves qualifier functions to utils and generates exponential backoff
