+++ engine/verification/requester/qualifier/incremental.go
@@ -1,8 +1,6 @@
 package qualifier
 
 import (
-	"time"
-
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/module/mempool"
 )
@@ -37,28 +35,4 @@ func (i ChunkDataRequestQualifier) CanDispatchRequest(chunkID flow.Identifier) b
 // increments the attempt field of request on its underlying mempool.
 func (i *ChunkDataRequestQualifier) OnRequestDispatched(chunkID flow.Identifier) bool {
 	return i.pendingRequests.IncrementAttempt(chunkID)
-}
-
-type RequestQualifierFunc func(uint64, time.Time, time.Duration) bool
-
-func UnlimitedAttemptQualifier() RequestQualifierFunc {
-	return func(uint64, time.Time, time.Duration) bool {
-		return true
-	}
-}
-
-func MaxAttemptQualifier(maxAttempts uint64) RequestQualifierFunc {
-	return func(attempts uint64, _ time.Time, _ time.Duration) bool {
-		return attempts < maxAttempts
-	}
-}
-
-func RetryAfterQualifier() RequestQualifierFunc {
-	return func(_ uint64, lastAttempt time.Time, retryAfter time.Duration) bool {
-		cutoff := lastAttempt.Add(retryAfter)
-		if cutoff.After(time.Now()) {
-			return true
-		}
-		return false
-	}
 }
